Move deps from workspace back into crates
This turns out to just be a mild hassle with little benefit.
There aren't many shared dependencies, and for the few
that we have, `cargo update` supports workspaces, so
keeping them up-to-date is not a problem.

The specific problem that I ran into is that cargo resolves
features once for the whole workspace instead of doing it
for each crate individually when running `cargo test` in the
root.